debian: fix that script runs with zsh (closes #1)

+++ debian_ubuntu/init.zsh
@@ -1,6 +1,7 @@
-#!/usr/bin/bash
+#!/usr/bin/zsh
 
-# EXECUTE THIS WITH BASH OR ZSH
+# EXECUTE THIS WITH ZSH ONLY!!
+
 # this creates simlinks into "~" to use the dotfiles.
 
 
@@ -18,27 +19,20 @@ ABS_DIR_PATH=$( cd "$(dirname "${BASH_SOURCE[0]}")" ; pwd -P )
 
 sh ~/dotfiles/common/install-zsh-users-plugins.sh
 
-echo "This program will create links for all in $ABS_DIR_PATH to $HOME."
+echo "This script will create links for all dotfiles in $ABS_DIR_PATH to $HOME."
 echo "Existing files will be deleted. Following files would be included:"
-echo "  .gitconfig-template => .gitconfig"
-echo "  .zshrc"
-echo "  .vimrc"
+echo "  .gitconfig-template => ~/.gitconfig"
+echo "  .zshrc => ~/.zshrc"
+echo "  .vimrc => ~/.vimrc"
 
-function onConfirmed {
+
+read "response?Are you sure ? [Y/n] "
+response=${response:l} #tolower
+if [[ $response =~ ^(yes|y| ) ]] || [[ -z $response ]]; then
     ln -fs $ABS_DIR_PATH/.zshrc ~
     ln -fs $ABS_DIR_PATH/.vimrc ~
     cp -f $ABS_DIR_PATH/.gitconfig-template ~/.gitconfig
-
-    echo "execute: 'source ~/.zshrc'"
-}
-
-while true; do
-    read -p "Do you wish to continue? [Y/n]" yn
-    case $yn in
-        [Yy]* ) onConfirmed; break;;
-        [Nn]* ) exit;;
-        * ) echo "Please answer yes or no.";;
-    esac
-done
-
+    
+    echo "execute: 'source ~/.zshrc'" 
+fi
